requirements.txt
absl-py==0.9.0
apache-beam==2.16.0
appnope==0.1.0
argh==0.26.2
astor==0.8.1
attrs==19.3.0
avro-python3==1.9.1
backcall==0.1.0
bleach==3.1.0
bumpversion==0.5.3
cachetools==3.1.1
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
configparser==4.0.2
coverage==5.0.1
crcmod==1.7
decorator==4.4.1
defusedxml==0.6.0
dill==0.3.0
docopt==0.6.2
entrypoints==0.3
fastavro==0.21.24
fasteners==0.15
filelock==3.0.12
future==0.18.2
gast==0.2.2
google-api-core==1.15.0
google-api-python-client==1.7.11
google-apitools==0.5.28
google-auth==1.10.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
google-cloud-bigquery==1.17.1
google-cloud-bigtable==1.0.0
google-cloud-core==1.1.0
google-cloud-datastore==1.7.4
google-cloud-pubsub==1.0.2
google-pasta==0.1.8
google-resumable-media==0.4.1
googleapis-common-protos==1.6.0
grpc-google-iam-v1==0.12.3
grpcio==1.26.0
h5py==2.10.0
hdfs==2.5.8
httplib2==0.12.0
idna==2.8
importlib-metadata==1.3.0
ipykernel==5.1.3
ipython==7.11.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.15.2
Jinja2==2.10.3
joblib==0.14.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
mistune==0.8.4
ml-metadata==0.15.2
mock==2.0.0
monotonic==1.5
more-itertools==8.0.2
nbconvert==5.6.1
nbformat==4.4.0
notebook==6.0.2
numpy==1.18.0
oauth2client==3.0.0
oauthlib==3.1.0
opt-einsum==3.1.0
packaging==19.2
pandas==0.25.3
pandocfilters==1.4.2
parso==0.5.2
pathtools==0.1.2
pbr==5.4.4
pep8==1.7.1
pexpect==4.7.0
pickleshare==0.7.5
pluggy==0.13.1
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.11.2
psutil==5.6.7
ptyprocess==0.6.0
py==1.8.1
pyarrow==0.14.1
pyasn1==0.4.8
pyasn1-modules==0.2.7
pydot==1.4.1
Pygments==2.5.2
pymongo==3.10.0
pyparsing==2.4.6
pyrsistent==0.15.6
pytest==5.3.2
pytest-cov==2.8.1
pytest-testmon==1.0.1
pytest-watch==4.2.0
python-dateutil==2.8.1
pytz==2019.3
PyYAML==3.13
pyzmq==18.1.1
qtconsole==4.6.0
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-learn==0.21.3
scipy==1.1.0
Send2Trash==1.5.0
six==1.13.0
tensorboard==2.0.2
tensorflow==2.0.0
tensorflow-data-validation==0.15.0
tensorflow-estimator==2.0.1
tensorflow-metadata==0.15.2
tensorflow-model-analysis==0.15.4
tensorflow-serving-api==2.0.0
tensorflow-transform==0.15.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tfx==0.15.0
tfx-bsl==0.15.3
toml==0.10.0
tornado==6.0.3
tox==3.14.3
traitlets==4.3.3
uritemplate==3.0.1
urllib3==1.25.7
virtualenv==16.7.9
watchdog==0.9.0
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wrapt==1.11.2
zipp==0.6.0
